requirements.txt
environs
flask
flask_migrate
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
markdown
jinja2
psycopg2
passlib
